Make singleton Class : PhotoManager
DataSource가 갖고있던, allPhotos를 singleton class를 만들어 책임을 분산.
이에 따른 코드 수정
